+++ twelvet-server/twelvet-server-system/src/main/java/com/twelvet/server/system/service/ISysRoleService.java
@@ -37,7 +37,7 @@ public interface ISysRoleService {
 	 * @param userId 用户ID
 	 * @return 选中角色ID列表
 	 */
-	List<Integer> selectRoleListByUserId(Long userId);
+	List<Long> selectRoleListByUserId(Long userId);
 
 	/**
 	 * 通过角色ID查询角色
